+++ android/src/test/kotlin/dev/teklund/nfc_wallet_suppression/NfcWalletSuppressionPluginTest.kt
@@ -14,14 +14,16 @@ import org.mockito.Mockito
  */
 
 internal class NfcWalletSuppressionPluginTest {
+
+  //TODO Fix tests
   @Test
-  fun onMethodCall_getPlatformVersion_returnsExpectedValue() {
+  fun onMethodCall_isSuppressed_returnsExpectedValue() {
     val plugin = NfcWalletSuppressionPlugin()
 
-    val call = MethodCall("getPlatformVersion", null)
+    val call = MethodCall("isSuppressed", null)
     val mockResult: MethodChannel.Result = Mockito.mock(MethodChannel.Result::class.java)
     plugin.onMethodCall(call, mockResult)
 
-    Mockito.verify(mockResult).success("Android " + android.os.Build.VERSION.RELEASE)
+    Mockito.verify(mockResult).success(false)
   }
 }