+++ example/integration_test/plugin_integration_test.dart
@@ -6,7 +6,6 @@
 // For more information about Flutter integration tests, please see
 // https://flutter.dev/to/integration-testing
 
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
@@ -15,11 +14,10 @@ import 'package:nfc_wallet_suppression/nfc_wallet_suppression.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  testWidgets('getPlatformVersion test', (WidgetTester tester) async {
-    final NfcWalletSuppression plugin = NfcWalletSuppression();
-    final String? version = await plugin.getPlatformVersion();
+  testWidgets('isSuppressed test', (WidgetTester tester) async {
+    final bool? suppressed = await NfcWalletSuppression.isSuppressed();
     // The version string depends on the host platform running the test, so
     // just assert that some non-empty string is returned.
-    expect(version?.isNotEmpty, true);
+    expect(suppressed, true);
   });
 }