+++ lib/nfc_wallet_suppression_platform_interface.dart
@@ -2,13 +2,19 @@ import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'nfc_wallet_suppression_method_channel.dart';
 
+/// An interface for a platform-specific implementation of the `nfc_wallet_suppression` plugin.
+///
+/// This class provides methods to request and release suppression of the NFC wallet.
+/// The platform-specific implementation should provide their own class that
+/// extends this one to be able to call the platform specific API.
 abstract class NfcWalletSuppressionPlatform extends PlatformInterface {
   /// Constructs a NfcWalletSuppressionPlatform.
   NfcWalletSuppressionPlatform() : super(token: _token);
 
   static final Object _token = Object();
 
-  static NfcWalletSuppressionPlatform _instance = MethodChannelNfcWalletSuppression();
+  static NfcWalletSuppressionPlatform _instance =
+      MethodChannelNfcWalletSuppression();
 
   /// The default instance of [NfcWalletSuppressionPlatform] to use.
   ///
@@ -23,18 +29,26 @@ abstract class NfcWalletSuppressionPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Future<String?> getPlatformVersion() {
-    throw UnimplementedError('platformVersion() has not been implemented.');
-  }
-
+  /// Request suppression of the NFC wallet.
+  ///
+  /// Returns `true` if the request was successful, `false` if it failed, and `null`
+  /// if the platform does not support this method.
   Future<bool?> requestSuppression() {
     throw UnimplementedError('requestSuppression() has not been implemented.');
   }
 
+  /// Release suppression of the NFC wallet.
+  ///
+  /// Returns `true` if the release was successful, `false` if it failed, and `null`
+  /// if the platform does not support this method.
   Future<bool?> releaseSuppression() {
     throw UnimplementedError('releaseSuppression() has not been implemented.');
   }
 
+  /// Check if the NFC wallet is currently suppressed.
+  ///
+  /// Returns `true` if the wallet is suppressed, `false` if it is not, and `null`
+  /// if the platform does not support this method.
   Future<bool?> isSuppressed() {
     throw UnimplementedError('isSuppressed() has not been implemented.');
   }