+++ test/nfc_wallet_suppression_test.dart
@@ -4,11 +4,10 @@ import 'package:nfc_wallet_suppression/nfc_wallet_suppression_method_channel.dar
 import 'package:nfc_wallet_suppression/nfc_wallet_suppression_platform_interface.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-class MockNfcWalletSuppressionPlatform with MockPlatformInterfaceMixin implements NfcWalletSuppressionPlatform {
+class MockNfcWalletSuppressionPlatform
+    with MockPlatformInterfaceMixin
+    implements NfcWalletSuppressionPlatform {
   bool _isSuppressing = false;
-
-  @override
-  Future<String?> getPlatformVersion() => Future.value('42');
 
   @override
   Future<bool?> requestSuppression() {
@@ -29,17 +28,36 @@ class MockNfcWalletSuppressionPlatform with MockPlatformInterfaceMixin implement
 }
 
 void main() {
-  final NfcWalletSuppressionPlatform initialPlatform = NfcWalletSuppressionPlatform.instance;
+  final NfcWalletSuppressionPlatform initialPlatform =
+      NfcWalletSuppressionPlatform.instance;
 
   test('$MethodChannelNfcWalletSuppression is the default instance', () {
     expect(initialPlatform, isInstanceOf<MethodChannelNfcWalletSuppression>());
   });
 
-  test('getPlatformVersion', () async {
-    NfcWalletSuppression nfcWalletSuppressionPlugin = NfcWalletSuppression();
-    MockNfcWalletSuppressionPlatform fakePlatform = MockNfcWalletSuppressionPlatform();
+  test('isSuppressed_defaultFalse', () async {
+    MockNfcWalletSuppressionPlatform fakePlatform =
+        MockNfcWalletSuppressionPlatform();
     NfcWalletSuppressionPlatform.instance = fakePlatform;
 
-    expect(await nfcWalletSuppressionPlugin.getPlatformVersion(), '42');
+    expect(await NfcWalletSuppression.isSuppressed(), isFalse);
+  });
+  test('isSuppressed_trueWhenSuppressed', () async {
+    MockNfcWalletSuppressionPlatform fakePlatform =
+        MockNfcWalletSuppressionPlatform();
+    NfcWalletSuppressionPlatform.instance = fakePlatform;
+
+    await NfcWalletSuppression.requestSuppression();
+    expect(await NfcWalletSuppression.isSuppressed(), isTrue);
+  });
+  test('isSuppressed_defaultFalse', () async {
+    MockNfcWalletSuppressionPlatform fakePlatform =
+        MockNfcWalletSuppressionPlatform();
+    NfcWalletSuppressionPlatform.instance = fakePlatform;
+
+    await NfcWalletSuppression.requestSuppression();
+    expect(await NfcWalletSuppression.isSuppressed(), isTrue);
+    await NfcWalletSuppression.releaseSuppression();
+    expect(await NfcWalletSuppression.isSuppressed(), isFalse);
   });
 }